+++ src/js/base/compat.js
@@ -89,6 +89,8 @@ if (
 ) {
     // eslint-disable-next-line no-var
     var fetch = null;
+    // eslint-disable-next-line no-var
+    var nodeFetch = null;
     if (typeof window !== "undefined" && typeof window.fetch !== "undefined") {
         fetch = window.fetch;
     } else if (typeof global !== "undefined" && typeof global.fetch !== "undefined") {
@@ -99,11 +101,33 @@ if (
         (typeof navigator === "undefined" || navigator.product !== "ReactNative")
     ) {
         fetch = require("node-fetch").default;
+        nodeFetch = fetch;
     } else if (typeof global !== "undefined" && typeof global.__VUE_SSR_CONTEXT__ !== "undefined") {
         // this is a workaround for Nuxt.js SSR built as standalone,
         // which does not have global.fetch populated
         fetch = require("node-fetch").default;
+        nodeFetch = fetch;
     }
+}
+
+if (nodeFetch) {
+    const http = require("http");
+    const https = require("https");
+    const process = require("process");
+    http.globalAgent.keepAlive = true;
+    http.globalAgent.keepAliveMsecs = 120000;
+    http.globalAgent.timeout = 60000;
+    http.globalAgent.scheduling = "fifo";
+    http.globalAgent.maxSockets = process.env.MAX_SOCKETS
+        ? parseInt(process.env.MAX_SOCKETS)
+        : Infinity;
+    https.globalAgent.keepAlive = true;
+    https.globalAgent.keepAliveMsecs = 120000;
+    https.globalAgent.timeout = 60000;
+    https.globalAgent.scheduling = "fifo";
+    https.globalAgent.maxSockets = process.env.MAX_SOCKETS
+        ? parseInt(process.env.MAX_SOCKETS)
+        : Infinity;
 }
 
 if (typeof module !== "undefined") {