+++ CHANGELOG.md
@@ -17,7 +17,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-*
+* Initials builder initialization when `useInitialsBuilderLogic` was `false`.
+
+## [2.0.0] - 2021-06-01
+
+### Added
+
+* [BREAKING CHANGES] Support for dynamic `initialsBuilder` logic sourced from server side 3DB ⚠️ (requires `initialsBuilder` signature change)
+
+## [1.26.0] - 2021-06-01
+
+### Added
+
+* `setTracking` method to the `Order` entity
+* `setMeta` method allowing dynamic set of metadata attributes for the `Order` entity
+* Method for generating an image URL from a query (`_queryToImageUrl`)
+
+## [1.25.9] - 2021-05-24
+
+### Added
+
+* Transport Rule API
+* `getTransportOrder` method
+* `attachmentOrder` method
+* `qualityAssureOrder` method
 
 ## [1.25.8] - 2021-05-14